Extract updateFlashcardSetting
